refactor: rename file_picker_android to android_file_picker
