+++ example/lib/src/picked_files_results.dart
@@ -1,5 +1,5 @@
 import 'package:file_picker/file_picker.dart';
-import 'package:file_picker_android/file_picker_android.dart';
+import 'package:android_file_picker/android_file_picker.dart';
 import 'package:flutter/material.dart';
 
 import 'file_picker_results.dart';
