+++ example/pubspec.yaml
@@ -14,7 +14,7 @@ dependencies:
   flutter:
     sdk: flutter
   file_picker: ^12.0.0-beta.8
-  file_picker_android: ^1.0.0
+  android_file_picker: ^1.0.0
   file_picker_darwin: ^1.0.0
   file_picker_linux: ^1.0.0
   file_picker_web: ^3.0.0
