+++ lib/src/file_picker.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:file_picker_android/file_picker_android.dart';
+import 'package:android_file_picker/android_file_picker.dart';
 import 'package:file_picker_platform_interface/file_picker_platform_interface.dart';
 import 'package:flutter/foundation.dart';
