+++ packages/android_file_picker/android/src/main/kotlin/com/mr/flutter/plugin/filepicker/FileInfo.kt
