+++ packages/android_file_picker/lib/android_file_picker.dart
@@ -1,0 +1,4 @@
+export 'src/android_platform_file.dart';
+export 'src/android_saf_handle.dart';
+export 'src/android_file_picker.dart';
+export 'src/android_file_picker_options.dart';
