+++ packages/android_file_picker/lib/src/android_platform_file.dart
