+++ packages/android_file_picker/lib/src/android_saf_handle.dart
@@ -1,7 +1,7 @@
 import 'package:file_picker_platform_interface/file_picker_platform_interface.dart';
 
-import 'file_picker_android.dart';
-import 'file_picker_android_options.dart';
+import 'android_file_picker.dart';
+import 'android_file_picker_options.dart';
 
 /// An opaque handle to an Android Storage Access Framework [uri].
 final class AndroidSAFHandle {
