+++ packages/android_file_picker/test/android_file_picker_test.dart
@@ -1,4 +1,4 @@
-import 'package:file_picker_android/file_picker_android.dart';
+import 'package:android_file_picker/android_file_picker.dart';
 import 'package:file_picker_platform_interface/file_picker_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
