--- packages/file_picker_android/lib/file_picker_android.dart
@@ -1,4 +1,0 @@
-export 'src/android_platform_file.dart';
-export 'src/android_saf_handle.dart';
-export 'src/file_picker_android.dart';
-export 'src/file_picker_android_options.dart';
