+++ packages/file_picker_platform_interface/README.md
@@ -2,7 +2,7 @@
 
 A common platform interface for the `file_picker` plugin.
 
-This package defines the interface that all platform implementation packages for `file_picker` (such as `file_picker_android`, `file_picker_darwin`, `file_picker_windows`, `file_picker_linux`, `file_picker_web`) must implement.
+This package defines the interface that all platform implementation packages for `file_picker` (such as `android_file_picker`, `file_picker_darwin`, `file_picker_windows`, `file_picker_linux`, `file_picker_web`) must implement.
 
 ## Usage
